Minor refactoring: consolidated 3 Youtube test classes into a single one with a file name according to the Effective Dart guidelines
