--- test/youtube/youtubeApiAdapter_getVideos_test.dart
@@ -1,23 +1,0 @@
-import 'package:influx/config.dart';
-import 'package:influx/utility/youtube/model/thumbnail_size.dart';
-import 'package:influx/utility/youtube/youtube_api_adapter.dart';
-import 'package:influx/utility/youtube/model/youtube_video_info.dart';
-import 'package:test_api/test_api.dart';
-import 'http_client_youtube_wbs_mock.dart';
-
-void main() {
-  test('fetching the last 20 video from wbs youtube channel', () async {
-    YoutubeApiAdapter youtubeApiAdapter = YoutubeApiAdapter();
-    List<YoutubeVideoInfo> videos = await youtubeApiAdapter.getVideos(
-        httpClient: HttpClientYoutubeWbsMock(),
-        apiKey: InFluxConfig.youtubeApiKey,
-        channelId: InFluxConfig.youtubeChannelId,
-        maxResults: 20);
-
-    assert(videos.length == 20);
-    assert(videos[0].id != null);
-    assert(videos[0].thumbnailUrls.containsKey(ThumbnailSize.SMALL));
-    assert(videos[0].publishedAt != null);
-    assert(videos[0].description != null);
-  });
-}
